+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Update the email notification template to include more check and last ping details
 - Improve the crontab snippet in the "Check Details" page (#465)
 - Add Signal integration (#428)
+- Change Zulip onboarding, ask for the zuliprc file (#202)
 
 ## Bug Fixes
 - Fix unwanted HTML escaping in SMS and WhatsApp notifications